% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{AddAvgCwtData}
\alias{AddAvgCwtData}
\title{Add averaged transfer function calculated from the CWT}
\usage{
AddAvgCwtData(framework, index)
}
\arguments{
\item{framework}{An analysis environment generated by \link[WaveletPhysiology]{BuildStructure}}

\item{index}{An index indicating which analysis slot needs to be used}
}
\value{
The original analysis environment, in which the averaged CWT transfer function has been added
        to a specific analysis slot.
}
\description{
Computes a vector of the frequency averaged transfer function values previously obtained from a
CWT transfer function in an analysis environment.
}
\examples{
# ADD EXAMPLE!
}
\author{
Alvaro Chao-Ecija
}
