% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWT_AlphaIndex.R
\name{AlphaIndexDWT}
\alias{AlphaIndexDWT}
\title{Discrete Wavelet Transform Transfer Function}
\usage{
AlphaIndexDWT(
  data,
  HF = 0.4,
  LF = 0.15,
  VLF = 0.04,
  wv = "d4",
  hrv = FALSE,
  error = 0.01,
  error.type = c("absolute", "relative")
)
}
\arguments{
\item{data}{A matrix with 3 columns containing time values (first column), RR and SBP values (second
and third column).}

\item{HF}{Maximum limit of the HF band. Default is 0.4 Hz}

\item{LF}{Maximum limit of the LF band. Default is 0.15 Hz}

\item{VLF}{Maximum limit of the VLF band. Default is 0.04 Hz}

\item{wv}{A wavelet for the analysis. Default is d4}

\item{hrv}{Boolean, should the heart rate variability results be returned too? Default is FALSE}

\item{error}{Tolerance for the PMODWPT. Default is 0.01}

\item{error.type}{Type of tolerance (absolute or relative) for the PMODWPT}
}
\value{
A list with the estimated components of the baroreflex transfer function in the wavelet domain:
\item{HF}{The computed vector of alpha indices for the HF band}
\item{LF}{The computed vector of alpha indices for the LF band}
\item{VLF}{The chosen maximum limit of the VLF band}
\item{Time}{The original vector of time values}
\item{type}{A character string specifying which type of BRS has been computed}
\item{hrv}{A list containing the variability results computed at the HF and LF bands, as well as the
           LF/HF ratio. Only returned if the hrv argument is set to TRUE}
}
\description{
Computes baroreflex sensitivity as a transfer function suing the Discrete Wavelet Transform
}
\details{
This function works in the following way: first the variables are adapted so that they can be
processed by the \href{https://CRAN.R-project.org/package=RHRV}{RHRV} package. This is done by creating structures similar to the one generated by
RHRV function \code{link[RHRV]{CreateHRVData}}, which are then filled by RR and BP data and can be recognized by RHRV. Then, this function sends the variables to the RHRV functions so that their variability
can be calculated using the Pruned-MODWT(PMODWT), an algorithm introduced in the RHRV package (see the reference
section for more details regarding package RHRV and the PMODWT algorithm or check the RHRV function \code{link[RHRV]{CalculatePowerBand}}). The obtained variability results are then used
to calculate a time-varying alpha index representative of the baroreflex sensitivity at the HF and LF bands.
Also, the function is able to return the computed heart variability results by \href{https://CRAN.R-project.org/package=RHRV}{RHRV} package.
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
AlphaIndex <- AlphaIndexDWT(Data, wv = "d8", error = 0.0005)
}
\references{
Rodriguez-Linares L, Mendez AJ, Lado MJ, Olivieri DN, Vila XA, Gomez-Conde I. An open source tool for heart rate variability spectral analysis. Comput Methods Programs Biomed. 2011;103(1):39-50.

Garcia Martinez CA, Otero Quintana A, Vila XA, Lado Tourino MJ, Rodriguez-Linares L, Rodriguez Presedo JM, Mendez Penin AJ. Heart Rate Variability Analysis with the R package RHRV. Springer Cham; 2017.

Constantino A. García, Abraham Otero, Xosé Vila, David G. Márquez. A new algorithm for wavelet-based heart rate variability analysis,
Biomedical Signal Processing and Control, Volume 8, Issue 6, 2013, Pages 542-550.
}
\author{
Alvaro Chao-Ecija
}
