% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesFunctions.R
\name{AnalyzeBRS}
\alias{AnalyzeBRS}
\title{Analyze baroreflex sensitivity}
\usage{
AnalyzeBRS(framework, locator, method = c("both", "dwt", "cwt"))
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator}{A locator index indicating which analysis slot needs to be used}

\item{method}{Which method to be used to estimate the transfer function. Default is both.}
}
\value{
The original analysis environment, in which baroreflex sensitivity has been estimated
        from the chosen data
}
\description{
Computes the baroreflex sensitivity from data located in an analysis environment.
}
\examples{
Data <- DataSimulation()
Study <- BuildStructure()
Study <- AddAnalysis(Study, name = "Simulation")
Study <- AddDataToAnalysis(Study, 1, Data$RR, Data$SBP, Data$Time, f = 1)
Study <- AnalyzeBRS(Study, 1)
}
\author{
Alvaro Chao-Ecija
}
