% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{AnalyzeBRSIndices}
\alias{AnalyzeBRSIndices}
\title{Analyze BRS indices}
\usage{
AnalyzeBRSIndices(
  framework,
  a_locator,
  t_locator,
  time_flags,
  method = c("both", "dwt", "cwt"),
  use.name = TRUE
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{a_locator}{An locator indicating which analysis slot needs to be used}

\item{t_locator}{An locator indicating which time inerval is to be used}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL.}

\item{method}{From which transfer function are the estimates to be computed, either cwt, dwt or both.
The default is both.}

\item{use.name}{Boolean, should the names of the subjects be used?}
}
\value{
The original framework with new computed indices.
}
\description{
Compute BRS indices from specific time intervals for all subjects.
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
Study <- BuildStructure()
Study <- AddAnalysis(Study, name = "Simulation")
Study <- AddDataToAnalysis(Study, 1, Data$RR, Data$SBP, Data$Time)
Study <- AnalyzeBRS(Study, 1)
Study <- GetAvgCwtData(Study, 1)

Study <- AddTimeInterval(Study, "First interval")
Study <- AddTimeInterval(Study, "Second interval")
Study <- AddTimeInterval(Study, "Third interval")

Study <- AnalyzeBRSIndices(Study, 1, 1, c(1.7, 3.3))
Study <- AnalyzeBRSIndices(Study, 1, 2, c(10, 11.7))
Study <- AnalyzeBRSIndices(Study, 1, 3, c(35, 37))
}
\author{
Alvaro Chao-Ecija
}
