% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{AnalyzeTransferFun}
\alias{AnalyzeTransferFun}
\title{Analyze baroreflex sensitivity}
\usage{
AnalyzeTransferFun(framework, index, method = c("both", "dwt", "cwt"))
}
\arguments{
\item{framework}{An analysis environment generated by \link[WaveletPhysiology]{BuildStructure}}

\item{index}{An index indicating which analysis slot needs to be used}

\item{method}{Which method to be used to estimate the transfer function. Default is both.}
}
\value{
The original analysis environment, in which baroreflex sensitivity has been estimated
        from the chosen data
}
\description{
Computes the baroreflex transfer function from data located in an analysis environment.
}
\examples{
# ADD EXAMPLE!
}
\author{
Alvaro Chao-Ecija
}
