% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesFunctions.R
\name{AssembleCwtBRS}
\alias{AssembleCwtBRS}
\title{Assemble BRS estimates computed with the CWT}
\usage{
AssembleCwtBRS(framework, locator)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator}{A locator indicating which analysis should be used}
}
\value{
A list containing the assembled transfer function or alpha index
}
\description{
Assembles a transfer function or alpha index computed with the CWT for further processing
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
Study <- BuildStructure()
Study <- AddAnalysis(Study, name = "Simulation")
Study <- AddDataToAnalysis(Study, 1, Data$RR, Data$SBP, Data$Time)
Study <- AnalyzeBRS(Study, 1)

TransferFun <- AssembleCwtBRS(Study, 1)
}
\author{
Alvaro Chao-Ecija
}
