% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaroWavelet-package.R
\docType{package}
\name{BaroWavelet-package}
\alias{BaroWavelet}
\alias{BaroWavelet-package}
\title{BaroWavelet: Baroreflex Sensitivity Assessment with Wavelet Transforms}
\description{
This package allows autonomic evaluation and baroreflex sensitivity assessment of interbeat intervals and blood pressure recordings using wavelet transforms. It includes different analysis techniques based on the continuous and discrete wavelet transforms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CIMES-USNA-UMA/BaroWavelet}
  \item Report bugs at \url{https://github.com/CIMES-USNA-UMA/BaroWavelet/issues}
}

}
\keyword{internal}
