% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Simulator from De Boer et Al.R
\name{DataSimulation}
\alias{DataSimulation}
\title{Simulation of cardiovascular data}
\usage{
DataSimulation(
  use.noise = TRUE,
  seed = 1,
  v = 9,
  s = c(0.25, 1.5, 2.75, 2, 0.75)
)
}
\arguments{
\item{use.noise}{Boolean. Add gaussian noise to the simulation}

\item{seed}{Constant for reproducibility purposes}

\item{v}{Vagal (instantaneous) component of the simulation. Default is 9}

\item{s}{Vector of sympathetic (lagged) components of the simulation}
}
\value{
A data frame containing timestamps and simulated RR and SBP recordings
}
\description{
Data simulation with different baroreflex behaviors. Based on the algorithm described by
de Boer and Karemaker (please consult the "references" section for more information)
}
\examples{
Data <- DataSimulation()
}
\references{
de Boer RW, Karemaker JM. Cross-Wavelet Time-Frequency Analysis Reveals Sympathetic Contribution to Baroreflex Sensitivity as Cause of Variable Phase Delay Between Blood Pressure and Heart Rate. Front Neurosci. 2019 Jul 9;13:694. Available from: https://www.frontiersin.org/articles/10.3389/fnins.2019.00694/full
}
\author{
Alvaro Chao-Ecija
}
