% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expected Values.R
\name{ExpectedValues}
\alias{ExpectedValues}
\title{Calculate expected values}
\usage{
ExpectedValues(
  fun,
  time_flags = NULL,
  weight = FALSE,
  thr = 0.5,
  use.coherence = TRUE
)
}
\arguments{
\item{fun}{A transfer function obtained either by \link[WaveletPhysiology]{TransferFunDWT} or
\link[WaveletPhysiology]{TransferFunCWT}}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL.}

\item{weight}{Boolean, should results be weighted by a Gaussian function? Default is FALSE.}

\item{thr}{Coherence threshold to ensure the reliability of the estimates. Default is 0.5}

\item{use.coherence}{Boolean, should a coherence threshold be used? Default is TRUE}
}
\value{
A vector containing the computed expected values at the HF and LF bands.
}
\description{
Computes the expected values of a transfer function
}
\examples{
# ADD EXAMPLE!
}
\references{
Violeta I. McLoone, John Ringwood. A system identification approach to baroreflex sensitivity
estimation. ISSC 2012; 2012.p. 1-6.
}
\author{
Alvaro Chao-Ecija
}
