% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{ExtractDataFromAnalysis}
\alias{ExtractDataFromAnalysis}
\title{Extract data from a specific analysis}
\usage{
ExtractDataFromAnalysis(framework, locator)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator}{A locator indicating which analysis should be used}
}
\value{
A list containg relevant data from a specific analysis
}
\description{
Extracts relevant data from a particular analysis
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
Study <- BuildStructure()
Study <- AddAnalysis(Study, name = "Simulation")
Study <- AddDataToAnalysis(Study, 1, Data$RR, Data$SBP, Data$Time)

New_data <- ExtractDataFromAnalysis(Data, 1)


PlotHRVTestResults(Study, 1)
}
\author{
Alvaro Chao-Ecija
}
