% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{GetExpectedValues}
\alias{GetExpectedValues}
\title{Get expected values}
\usage{
GetExpectedValues(
  framework,
  a_index,
  t_index,
  time_flags,
  method = c("both", "dwt", "cwt"),
  use.name = TRUE
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[WaveletPhysiology]{BuildStructure}}

\item{a_index}{An index indicating which analysis slot needs to be used}

\item{t_index}{An index indicating which time inerval is to be used}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL.}

\item{method}{From which transfer function are the estimates to be computed, either cwt, dwt or both.
The default is both.}

\item{use_name}{Boolean, should the names of the subjects be used?}
}
\value{
The original framework with new computed estimates.
}
\description{
Compute expected values from specific time intervals.
}
\examples{
# ADD EXAMPLE!
}
\author{
Alvaro Chao-Ecija
}
