% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Indices.R
\name{IndividualIndices}
\alias{IndividualIndices}
\title{Calculate individual indices}
\usage{
IndividualIndices(
  fun,
  time_flags = NULL,
  thr = 0.5,
  use.coherence = TRUE,
  method = c("median", "mean"),
  use.phase = FALSE
)
}
\arguments{
\item{fun}{BRS obtained either by \link[BaroWavelet]{AlphaIndexDWT} or
\link[BaroWavelet]{TransferFunCWT}}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL.}

\item{thr}{Coherence threshold to ensure the reliability of the estimates. Default is 0.5}

\item{use.coherence}{Boolean, should a coherence threshold be used? Default is TRUE}

\item{method}{Which method, either the mean or the median, should be used to obtain the indices}

\item{use.phase}{Boolean, should indices be computed from CWT phase difference estimates? Default is FALSE}
}
\value{
A matrix containing data (either medians and IQRs, or means and standard deviations)
identifying indices at the HF and LF bands for the specific subjects.
}
\description{
Computes an individual BRS index from a single subject
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
AlphaIndex <- AlphaIndexDWT(Data, wv = "d8", error = 0.0005)

IndividualIndices(AlphaIndex, c(0, 100/60))
}
\references{
Lazaro J, Gil E, Orini M, Laguna P, Bailón R. Baroreflex Sensitivity Measured by Pulse Photoplethysmography. Front Neurosci. 2019 Apr 18;13:339. Available from: https://www.frontiersin.org/articles/10.3389/fnins.2019.00339/full
}
\author{
Alvaro Chao-Ecija
}
