% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpolateData.R
\name{InterpolateData}
\alias{InterpolateData}
\title{Interpolate data}
\usage{
InterpolateData(x, f = 4)
}
\arguments{
\item{x}{Multivariate time series to be interpolated}

\item{f}{Sample frequency. Default is 4 Hz}
}
\value{
An interpolated time series.
}
\description{
Interpolates data up to a certain sample frequency
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)

}
\author{
Alvaro Chao-Ecija
}
