% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{ModelClinicalData}
\alias{ModelClinicalData}
\title{Model clinical data}
\usage{
ModelClinicalData(
  framework,
  type = c("BRS", "HRV"),
  band = c("HF", "LF", "LFHF"),
  segment = 1,
  variable = 1,
  use.names = TRUE
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{type}{Which data to be used, either HRV or BRS data}

\item{band}{Indicates if the HF or LF band results (or the LF/HF ratio) are included in the model}

\item{segment}{Indicates a specific time interval}

\item{variable}{Indicates a specific clinical variable}

\item{use.names}{Boolean, should the names of the subjects be used?}
}
\value{
The results of a linear regression model with the chosen data
}
\description{
Allows the modelling of external clinical data alongside the BRS and HRV data
}
\author{
Alvaro Chao-Ecija
}
