% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{PlotAnalyzedHRV}
\alias{PlotAnalyzedHRV}
\title{Plot analyzed heart rate variability}
\usage{
PlotAnalyzedHRV(
  framework,
  locator,
  time_col = "brown",
  HFcolor = "yellow",
  LFcolor = "green",
  time_flags = NULL,
  tem = FALSE,
  plotHF = TRUE,
  plotLF = TRUE,
  newPlot = TRUE
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator}{A locator indicating which analysis slot needs to be used}

\item{time_col}{Color used to highlight a specific time interval. Default is brown}

\item{HFcolor}{Color to be used to highlight the HF band}

\item{LFcolor}{Color to be used to highlight the LF band}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL.}

\item{tem}{Boolean, creates a temporal file for the plot. Default is FALSE}

\item{plotHF}{Boolean, plot results form the HF band. Default is TRUE}

\item{plotLF}{Boolean, plot results from the LF band. Default is TRUE}

\item{newPlot}{Boolean, generates a new plot without overwriting a previous plot. Default is TRUE}
}
\value{
None
}
\description{
Plots the HRV associated to a specific BRS estimation
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
Study <- BuildStructure()
Study <- AddAnalysis(Study, name = "Simulation")
Study <- AddDataToAnalysis(Study, 1, Data$RR, Data$SBP, Data$Time)
Study <- AnalyzeBRS(Study, 1)
Study <- GetAvgCwtData(Study, 1)


PlotAnalyzedHRV(Study, 1)
}
\author{
Alvaro Chao-Ecija
}
