% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{PlotTestResults}
\alias{PlotTestResults}
\title{Plot results from group testing of BRS}
\usage{
PlotTestResults(
  framework,
  locator,
  tem = FALSE,
  newPlot = FALSE,
  draw_paired = FALSE
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator}{A locator indicating which test should be used}

\item{tem}{Boolean, creates a temporal file for the plot. Default is FALSE}

\item{newPlot}{Boolean, generates a new plot without overwriting a previous plot. Default is TRUEhould the HRV be normalized before comparison? Default is FALSE}

\item{draw_paired}{Boolean. If the groups are paired, draw lines connecting each couple of paired estimates. Default is FALSE}
}
\value{
None
}
\description{
Plots the results obtained from comparing BRS indices between two groups
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
Study <- BuildStructure(error = 0.01)
Study <- AddAnalysis(Study, name = "Simulation 1")
Study <- AddAnalysis(Study, name = "Simulation 2")
Study <- AddAnalysis(Study, name = "Simulation 3")
Study <- AddAnalysis(Study, name = "Simulation 4")
Study <- AddDataToAnalysis(Study, 1, Data$RR[1:300], Data$SBP[1:300], Data$Time[1:300])
Study <- AddDataToAnalysis(Study, 2, Data$RR[501:800], Data$SBP[501:800], Data$Time[1:300])
Study <- AddDataToAnalysis(Study, 3, Data$RR[1001:1300], Data$SBP[1001:1300], Data$Time[1:300])
Study <- AddDataToAnalysis(Study, 4, Data$RR[1501:1800], Data$SBP[1501:1800], Data$Time[1:300])
for(n in 1:4) Study <- AnalyzeBRS(Study, n)
for(n in 1:4) Study <- AddAvgCwtData(Study, 1)

Study <- AddTimeInterval(Study, "First interval")
Study <- AddTimeInterval(Study, "Second interval")

Study <- AnalyzeBRSIndices(Study, 1, 1, c(0, 2.5))
Study <- AnalyzeBRSIndices(Study, 1, 2, c(2.6, 3))
Study <- AnalyzeBRSIndices(Study, 2, 1, c(0, 2.5))
Study <- AnalyzeBRSIndices(Study, 2, 2, c(2.6, 3))
Study <- AnalyzeBRSIndices(Study, 3, 1, c(0, 2.5))
Study <- AnalyzeBRSIndices(Study, 3, 2, c(2.6, 3))
Study <- AnalyzeBRSIndices(Study, 4, 1, c(0, 2.5))
Study <- AnalyzeBRSIndices(Study, 4, 2, c(2.6, 3))

Study <- TestGroups(Study, 1, 2, "Example test")

PlotTestResults(Study, 1)
}
\author{
Alvaro Chao-Ecija
}
