% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing with RHRV.R
\name{PrepareTachoForRHRV}
\alias{PrepareTachoForRHRV}
\title{Prepare tachogram for RHRV}
\usage{
PrepareTachoForRHRV(time)
}
\arguments{
\item{time}{Time support for the HR recording}
}
\description{
Prepare tachogram for RHRV
}
\author{
Alvaro Chao-Ecija
}
