% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing_with_RHRV.R
\name{PreprocessData}
\alias{PreprocessData}
\title{Preprocess data with RHRV}
\usage{
PreprocessData(
  data,
  maxSBP = 300,
  minSBP = 30,
  minbpm = 25,
  maxbpm = 200,
  use.RHRV = TRUE
)
}
\arguments{
\item{data}{Cardiovascular data to be processed}

\item{maxSBP}{Maximum value allowed for the SBP}

\item{minSBP}{Minimum value allowed for the SBP}

\item{minbpm}{Minimum bpm allowed for HR}

\item{maxbpm}{Maximum bpm allowed for HR}

\item{use.RHRV}{Boolean. Should RHRV be used in the preprocessing step? Default is TRUE.}
}
\description{
Preprocess data with RHRV
}
\details{
This function serves as a proposal for the preprocessing of raw data. It thresholds the
data according to pressure levels that can be defined by the user, as well as heart rate levels.
The heart rate threshold levels are based on the work of AUTHOR et al. (please check the reference
section for more information). Alternatively, it can send data to RHRV so that the RHRV package
will filter the RR intervals. The function creates a structure similar to the one generated by
\code{link[RHRV]{CreateHRVData}} and fills it with a time support, that will be processed by
RHRV functions \code{link[RHRV]{BuildNIHR}} and \code{link[RHRV]{FilterNIHR}}.
}
\references{
Garcia Martinez CA, Otero Quintana A, Vila XA, Lado Tourino MJ, Rodriguez-Linares L, Rodriguez Presedo JM, Mendez Penin AJ. Heart Rate Variability Analysis with the R package RHRV. Springer Cham; 2017.
}
\author{
Alvaro Chao-Ecija
}
