% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWT TransferFun.R
\name{SendDataToRHRV}
\alias{SendDataToRHRV}
\title{Send compatible objects to RHRV}
\usage{
SendDataToRHRV(RHRVobjects, HF = 0.4, LF = 0.15, VLF = 0.04, wv = "d4")
}
\arguments{
\item{RHRVobjects}{A list of objects generated by \link{WaveletPhysiology}{GenerateRHRVObjects}}
}
\value{
A list with two objects, containing the results from the analyses performed by RHRV.
}
\description{
Allows compatible objects to be analyzed by the RHRV package.
}
\examples{
# ADD EXAMPLE!
}
\author{
Alvaro Chao-Ecija
}
\keyword{internal}
