% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{ShowIndexes}
\alias{ShowIndexes}
\title{Show indexes}
\usage{
ShowIndexes(framework, type = c("analyses", "controls", "intervals", "tests"))
}
\arguments{
\item{framework}{An analysis environment generated by \link[WaveletPhysiology]{BuildStructure}}

\item{type}{Type of index to be shown}
}
\value{
A matrix containing the indexes to be shown
}
\description{
Show indexes regarding a specific analysis environment.
}
\examples{
# ADD EXAMPLE!
}
\author{
Alvaro Chao-Ecija
}
