% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{ShowLocatorIndices}
\alias{ShowLocatorIndices}
\title{Show locator indices}
\usage{
ShowLocatorIndices(
  framework,
  type = c("analyses", "controls", "intervals", "tests")
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{type}{Type of locator index to be shown}
}
\value{
A matrix containing the indexes to be shown
}
\description{
Show locator indices regarding a specific analysis environment.
}
\examples{
Study <- BuildStructure()
ShowLocatorIndices(Study, "analyses")
ShowLocatorIndices(Study, "controls")
ShowLocatorIndices(Study, "intervals")
ShowLocatorIndices(Study, "tests")
}
\author{
Alvaro Chao-Ecija
}
