% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT TransferFun.R
\name{SmoothTransforms}
\alias{SmoothTransforms}
\title{Smooth Continuous Wavelet Transforms results}
\usage{
SmoothTransforms(x, y, chosen.dj = 1/20)
}
\arguments{
\item{x}{The Continuous Wavelet Transform of variable x}

\item{y}{The Continuous Wavelet Transform of variable y}

\item{chosen.dj}{Frequency resolution. Default is 1/20}
}
\value{
A list with the smoothed transforms:
\item{sWT.x}{The smoothed transform of variable x}
\item{sWT.y}{The smoothed transform of variable y}
\item{sXWT}{The smoothed cross-wavelet transform between x and y}
}
\description{
Smooths results obtained for Continuous Wavelet Transform, so as to compute the
baroreflex transfer function.
}
\details{
This function is based on the code that function biwavelet uses to calculate the
wavelet coherence. For more information, please check the reference section.
}
\examples{
# ADD EXAMPLE!
}
\author{
Alvaro Chao-Ecija
}
\keyword{internal}
