% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesFunctions.R
\name{SplitByCoherence}
\alias{SplitByCoherence}
\title{Split CWT BRS with coherence criteria}
\usage{
SplitByCoherence(
  fun,
  thr = 0.5,
  use.thr = TRUE,
  time_flags = NULL,
  use.phase = FALSE
)
}
\arguments{
\item{fun}{BRS obtained from \link[BaroWavelet]{TransferFunCWT}}

\item{thr}{Coherence threshold to ensure the reliability of the estimates. Default is 0.5}

\item{use.thr}{Boolean, should a coherence threshold be used? Default is TRUE}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL}

\item{use.phase}{Use phase estimates instead of gains. Default is FALSE}
}
\value{
A list containg the assembled transfer function or alpha index
}
\description{
Splits BRS estimates computed with the CWT into the HF and LF bands after applying a coherence thresholding
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
BRS <- TransferFunCWT(Data)
Split_BRS <- SplitByCoherence(BRS)
}
\author{
Alvaro Chao-Ecija
}
