% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesFunctions.R
\name{TestHRV}
\alias{TestHRV}
\title{Test and compare HRV data organized in groups}
\usage{
TestHRV(
  framework,
  locator1,
  locator2,
  name = NULL,
  method = NULL,
  paired = NULL,
  use.names = TRUE,
  sig = 0.05,
  normalize = FALSE,
  showerror = TRUE
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator1}{A locator index indicating which interval slot needs to be used}

\item{locator2}{A locator index indicating another interval slot to perform the comparison}

\item{name}{An optional name for a specific comparison}

\item{method}{Which test should be used to perform the comparison. Default is NULL (it will be chosen automatically)}

\item{paired}{Boolean. Are the two groups paired? Default is NULL}

\item{use.names}{Boolean, should the names of the subjects be used?}

\item{sig}{Value indicating the chosen alpha error for the comparison. Default is 0.05}

\item{normalize}{Boolean. Should the HRV be normalized before comparison? Default is FALSE}

\item{showerror}{Boolean. Should the function be stopped and an error message be shown if an error is detected?}
}
\value{
The original framework with the test results attached
}
\description{
Compares two groups of paired or umpared HRV data to check for significant differences
}
\examples{
niData <- DataSimulation()
Data <- InterpolateData(niData, f = 1)
Study <- BuildStructure(error = 0.01)
Study <- AddAnalysis(Study, name = "Simulation 1")
Study <- AddAnalysis(Study, name = "Simulation 2")
Study <- AddAnalysis(Study, name = "Simulation 3")
Study <- AddAnalysis(Study, name = "Simulation 4")

Study <- AddDataToAnalysis(Study, 1, Data$RR[1:300], Data$SBP[1:300], Data$Time[1:300])
Study <- AddDataToAnalysis(Study, 2, Data$RR[501:800], Data$SBP[501:800], Data$Time[1:300])
Study <- AddDataToAnalysis(Study, 3, Data$RR[1001:1300], Data$SBP[1001:1300], Data$Time[1:300])
Study <- AddDataToAnalysis(Study, 4, Data$RR[1501:1800], Data$SBP[1501:1800], Data$Time[1:300])

Study <- AddDataToAnalysis(Study, 1, niData$RR[1:300], niData$SBP[1:300], 
niData$Time[1:300], raw = TRUE)
Study <- AddDataToAnalysis(Study, 2, niData$RR[501:800], niData$SBP[501:800], 
niData$Time[1:300], raw = TRUE)
Study <- AddDataToAnalysis(Study, 3, niData$RR[1001:1300], niData$SBP[1001:1300], 
niData$Time[1:300], raw = TRUE)
Study <- AddDataToAnalysis(Study, 4, niData$RR[1501:1800], niData$SBP[1501:1800], 
niData$Time[1:300], raw = TRUE)

for(n in 1:4) Study <- AnalyzeBRS(Study, n)
for(n in 1:4) Study <- AddAvgCwtData(Study, 1)

Study <- AddTimeInterval(Study, "First interval")
Study <- AddTimeInterval(Study, "Second interval")

Study <- AnalyzeBRSIndices(Study, 1, 1, c(0, 2.5))
Study <- AnalyzeBRSIndices(Study, 1, 2, c(2.6, 3))
Study <- AnalyzeBRSIndices(Study, 2, 1, c(0, 2.5))
Study <- AnalyzeBRSIndices(Study, 2, 2, c(2.6, 3))
Study <- AnalyzeBRSIndices(Study, 3, 1, c(0, 2.5))
Study <- AnalyzeBRSIndices(Study, 3, 2, c(2.6, 3))
Study <- AnalyzeBRSIndices(Study, 4, 1, c(0, 2.5))
Study <- AnalyzeBRSIndices(Study, 4, 2, c(2.6, 3))

Study <- TestHRV(Study, 1, 2, "Example test")
}
\author{
Alvaro Chao-Ecija
}
