% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_Individual_Variables.R
\name{TestIndHRandBP}
\alias{TestIndHRandBP}
\title{Test Individual HR and BP}
\usage{
TestIndHRandBP(data, time_flags1, time_flags2)
}
\arguments{
\item{data}{A matrix with HR and BP recordings, as well as timestamps}

\item{time_flags1}{A vector containing the minimum and maximum limits for the first time interval, in minutes}

\item{time_flags2}{A vector containing the minimum and maximum limits for the second time interval, in minutes}
}
\value{
A vector contaning p-values computed for the HR and the SBP
}
\description{
Significant test to compare two time-domain segments in HR and BP recordings
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)

TestIndHRandBP(Data, c(0, 1.7), c(35, 38))
TestIndHRandBP(Data, c(0, 1.7), c(8, 9.5))
}
\author{
Alvaro Chao-Ecija
}
