% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test Individual Variables.R
\name{TestIndHRV}
\alias{TestIndHRV}
\title{Test Individual HRV}
\usage{
TestIndHRV(fun, time_flags1, time_flags2)
}
\arguments{
\item{fun}{BRS obtainedfrom \link[BaroWavelet]{AlphaIndexDWT}, with HRV data.}

\item{time_flags1}{A vector containing the minimum and maximum limits for the first time interval, in minutes.}

\item{time_flags2}{A vector containing the minimum and maximum limits for the second time interval, in minutes.}
}
\value{
A vector contaning p-values computed for the HF and LF bands, as well as the LF/HF ratio
}
\description{
Significant test to compare two time-domain segments from an individual HRV computation
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
AlphaIndex <- AlphaIndexDWT(Data, wv = "d8" error = 0.0005, hrv = TRUE)

TestIndHRV(AlphaIndex, c(0, 1.7), c(35, 38))
TestIndHRV(AlphaIndex, c(0, 1.7), c(8, 9.5))
}
\author{
Alvaro Chao-Ecija
}
