% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Indices.R
\name{TimeDomainValues}
\alias{TimeDomainValues}
\title{Calculate time domain values}
\usage{
TimeDomainValues(data, time_flags = NULL, method = c("median", "mean"))
}
\arguments{
\item{data}{matrix with 3 columns containing time values (first column), RR and SBP values (second
and third column), or a data frame containing these variables.}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL.}

\item{method}{Which method, either the mean or the median, should be used to obtain the indices}
}
\value{
A list with HR and BP levels
}
\description{
Computes HR and BP time domain indices
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)

TimeDomainValues(Data, c(0, 100/60), "mean")
}
\author{
Alvaro Chao-Ecija
}
