% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWT_BRS_Methods.R
\name{TransferFunCWT}
\alias{TransferFunCWT}
\title{Continuous Wavelet Transform Transfer Function}
\usage{
TransferFunCWT(
  data,
  HF = 0.4,
  LF = 0.15,
  VLF = 0.04,
  chosen.dj = 1/20,
  demean = TRUE,
  smooth = TRUE,
  alpha = FALSE
)
}
\arguments{
\item{data}{A matrix with 3 columns containing time values (first column), RR and SBP values (second
and third column).}

\item{HF}{Maximum limit of the HF band. Default is 0.4 Hz}

\item{LF}{Maximum limit of the LF band. Default is 0.15 Hz}

\item{VLF}{Maximum limit of the VLF band. Default is 0.04 Hz}

\item{chosen.dj}{Frequency resolution to be passed to the \code{link[biwavelet]{wt}} function. Default is 1/20}

\item{demean}{Boolean, should the data be demeaned before analysis? Default is TRUE}

\item{smooth}{Boolean, smooth the transforms so that the BRS correlates with the coherence. Default is TRUE}

\item{alpha}{Boolean, compute an alpha index representative of the BRS instead of the transfer function. Default is FALSE}
}
\value{
A list with the estimated components of the baroreflex transfer function in the wavelet domain:
\item{TransferFun}{The computed baroreflex transfer function}
\item{Coherence}{The computed coherence between the two variables}
\item{Freqs}{A vector of frequencies for which the transfer function has been computed}
\item{Cone}{The computed cone of influence for the Continuous Wavelet Transform}
\item{Time}{The original vector of time values}
\item{HF}{The chosen maximum limit of the HF band}
\item{LF}{The chosen maximum limit of the LF band}
\item{VLF}{The chosen maximum limit of the VLF band}
\item{type}{A character string specifying which type of BRS has been computed}
\item{Cospectrum}{Computed cospectrum}
\item{Quadrature}{Computed quadrature}
\item{Phase}{Computed phase difference}
\item{Scales}{Vector of scales used in the transform}
\item{Alpha}{Boolean value indicating if the computed BRS is an alpha index}
}
\description{
Computes baroreflex sensitivity as a transfer function using the Continuous Wavelet Transform
}
\details{
This function makes use of the Continuous Wavelet Transform methods
provided by package \href{https://CRAN.R-project.org/package=biwavelet}{biwavelet} to
compute the baroreflex sensitivity. It employs the functions \code{link[biwavelet]{wt}}
and \code{link[biwavelet]{smooth.wavelet}}. This last function is used in a smoothing routine
based on the one that biwavelet function \code{link[biwavelet]{wtc}} uses to smooth
the wavelet transforms. Further information regarding this method of computation, as
well as the \href{https://CRAN.R-project.org/package=biwavelet}{biwavelet} package,
can be accessed through the references section.
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
TransferFun <- TransferFunCWT(Data)
}
\references{
Keissar K, Maestri R, Pinna GD, La Rovere MT, Gilad O. Non-invasive
baroreflex sensitivity assessment using wavelet transfer function-based
time-frequency analysis. Physiol Meas. 2010 ;31(7):1021-36.

Tarik C. Gouhier, Aslak Grinsted, Viliam Simko (2021). R package biwavelet:
Conduct Univariate and Bivariate Wavelet Analyses (Version 0.20.21).
Available from https://github.com/tgouhier/biwavelet

A Chao-Ecija , MS Dawid-Milner , BaroWavelet: An R-based tool for dynamic 
baroreflex evaluation through wavelet analysis techniques, Computer Methods 
and Programs in Biomedicine (2023), doi: https://doi.org/10.1016/j.cmpb.2023.107758
}
\author{
Alvaro Chao-Ecija
}
