% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyses Functions.R
\name{AddDataToAnalysis}
\alias{AddDataToAnalysis}
\title{Add data to analysis}
\usage{
AddDataToAnalysis(framework, index, RR, SBP, time)
}
\arguments{
\item{framework}{An analysis environment generated by \link[WaveletPhysiology]{BuildStructure}}

\item{index}{An index indicating which analysis slot needs to be used}

\item{RR}{A vector with RR data}

\item{SBP}{A vector with systolic blood pressure data}

\item{time}{A vector of time values}
}
\value{
The original analysis environment, in which new data has been placed in the chosen
         analysis slot.
}
\description{
Add data to a previously generated analysis slot.
}
\examples{
# ADD EXAMPLE!
}
\author{
Alvaro Chao-Ecija
}
