% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesFunctions.R
\name{AddDataToAnalysis}
\alias{AddDataToAnalysis}
\title{Add data to analysis}
\usage{
AddDataToAnalysis(framework, locator, RR, SBP, time)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator}{An locator indicating which analysis slot needs to be used}

\item{RR}{A vector with RR data}

\item{SBP}{A vector with systolic blood pressure data}

\item{time}{A vector of time values}
}
\value{
The original analysis environment, in which new data has been placed in the chosen
         analysis slot.
}
\description{
Add data to a previously generated analysis slot.
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
Study <- BuildStructure()
Study <- AddAnalysis(Study, name = "Simulation")
Study <- AddDataToAnalysis(Study, 1, Data$RR, Data$SBP, Data$Time)
}
\author{
Alvaro Chao-Ecija
}
