% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesFunctions.R
\name{AnalyzeBRSIndices}
\alias{AnalyzeBRSIndices}
\title{Analyze BRS indices}
\usage{
AnalyzeBRSIndices(
  framework,
  locator_a,
  locator_t,
  time_flags,
  method = c("both", "dwt", "cwt"),
  use.name = TRUE,
  use.raw = TRUE
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator_a}{A locator index indicating which analysis slot needs to be used}

\item{locator_t}{A locator index indicating which time interval is to be used}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL.}

\item{method}{From which transfer function are the estimates to be computed, either cwt, dwt or both.
The default is both.}

\item{use.name}{Boolean, should the names of the subjects be used?}

\item{use.raw}{Boolean. If TRUE, HR and BP estimates will be computed from non-interpolated time series}
}
\value{
The original framework with new computed indices.
}
\description{
Compute BRS indices from specific time intervals for all subjects.
}
\examples{
niData <- DataSimulation()
Data <- InterpolateData(niData, f = 1)
Study <- BuildStructure()
Study <- AddAnalysis(Study, name = "Simulation")
Study <- AddDataToAnalysis(Study, 1, Data$RR, Data$SBP, Data$Time)
Study <- AddDataToAnalysis(Study, 1, niData$RR, niData$SBP, niData$Time, raw = TRUE)
Study <- AnalyzeBRS(Study, 1)
Study <- AddAvgCwtData(Study, 1)

Study <- AddTimeInterval(Study, "First interval")
Study <- AddTimeInterval(Study, "Second interval")
Study <- AddTimeInterval(Study, "Third interval")

Study <- AnalyzeBRSIndices(Study, 1, 1, c(1.7, 3.3))
Study <- AnalyzeBRSIndices(Study, 1, 2, c(10, 11.7))
Study <- AnalyzeBRSIndices(Study, 1, 3, c(35, 37))
}
\author{
Alvaro Chao-Ecija
}
