% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaroWavelet-package.R
\docType{package}
\name{BaroWavelet-package}
\alias{BaroWavelet}
\alias{BaroWavelet-package}
\title{BaroWavelet: Baroreflex Sensitivity Assessment with Wavelet Transforms}
\description{
Performs autonomic evaluation and baroreflex sensitivity assessment of interbeat intervals and blood pressure recordings using wavelet transforms. It includes different analysis techniques based on the continuous and discrete wavelet transforms. For more details please check the following in-press article: A Chao-Ecija , MS Dawid-Milner , BaroWavelet: An R-based tool for dynamic baroreflex evaluation through wavelet analysis techniques, Computer Methods and Programs in Biomedicine (2023), \doi{10.1016/j.cmpb.2023.107758}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CIMES-USNA-UMA/BaroWavelet}
  \item Report bugs at \url{https://github.com/CIMES-USNA-UMA/BaroWavelet/issues}
}

}
\keyword{internal}
