% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesFunctions.R
\name{PlotAnalyzedBRS}
\alias{PlotAnalyzedBRS}
\title{Plot analyzed baroreflex sensitivity}
\usage{
PlotAnalyzedBRS(
  framework,
  locator,
  method = c("dwt", "cwt", "cwt.avg", "cwt.phase"),
  time_col = "brown",
  HFcolor = "yellow",
  LFcolor = "green",
  time_flags = NULL,
  use.coherence = TRUE,
  nfreqs = 7,
  tem = FALSE,
  newPlot = TRUE,
  plotHF = TRUE,
  plotLF = TRUE,
  thr = NULL,
  ylim = NULL,
  use.ggplot = TRUE,
  size.axis = 100,
  size.labels = 100,
  size.title = 100,
  d.labels = 100
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator}{A locator indicating which analysis slot needs to be used}

\item{method}{Which method to be used to plot the transfer function.}

\item{time_col}{Color used to highlight a specific time interval. Default is brown}

\item{HFcolor}{Color to be used to highlight the HF band for the average CWT plot}

\item{LFcolor}{Color to be used to highlight the LF band for the average CWT plot}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL.}

\item{use.coherence}{Boolean, should a coherence threshold be used in the analyses? Default is TRUE.}

\item{nfreqs}{Number of frequencies to be used. Default is 7}

\item{tem}{Boolean, creates a temporal file for the plot. Default is FALSE}

\item{newPlot}{Boolean, generates a new plot without overwriting a previous plot. Default is TRUE}

\item{plotHF}{Boolean, plot results form the HF band. Default is TRUE}

\item{plotLF}{Boolean, plot results from the LF band. Default is TRUE}

\item{thr}{Coherence threshold to be used for the plot. Default is NULL}

\item{ylim}{Maximum y axis limit. Default is NULL}

\item{use.ggplot}{Boolean, use methods from ggplot2 package to plot the results. Default is TRUE}

\item{size.axis}{Percentage of scaling of axis values. Default is 100}

\item{size.labels}{Percentage of scaling of axis labels. Default is 100}

\item{size.title}{Percentage of scaling of plot titles. Default is 100}

\item{d.labels}{Percentage of scaling of distance between axis and labels. Default is 100}
}
\value{
None
}
\description{
Plots an already computed transfer function from a specific analysis in an analysis environment.
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
Study <- BuildStructure()
Study <- AddAnalysis(Study, name = "Simulation")
Study <- AddDataToAnalysis(Study, 1, Data$RR, Data$SBP, Data$Time)
Study <- AnalyzeBRS(Study, 1)
Study <- AddAvgCwtData(Study, 1)


PlotAnalyzedBRS(Study, 1, "cwt", newPlot = FALSE)

}
\author{
Alvaro Chao-Ecija
}
