% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysesFunctions.R
\name{PlotAnalyzedHRV}
\alias{PlotAnalyzedHRV}
\title{Plot analyzed heart rate variability}
\usage{
PlotAnalyzedHRV(
  framework,
  locator,
  time_col = "brown",
  HFcolor = "yellow",
  LFcolor = "green",
  time_flags = NULL,
  tem = FALSE,
  plotHF = TRUE,
  plotLF = TRUE,
  ratio = TRUE,
  normalize = FALSE,
  newPlot = TRUE,
  ylim = NULL,
  use.ggplot = FALSE,
  size.axis = 100,
  size.labels = 100,
  size.title = 100,
  d.labels = 100
)
}
\arguments{
\item{framework}{An analysis environment generated by \link[BaroWavelet]{BuildStructure}}

\item{locator}{A locator index indicating which analysis slot needs to be used}

\item{time_col}{Color used to highlight a specific time interval. Default is brown}

\item{HFcolor}{Color to be used to highlight the HF band}

\item{LFcolor}{Color to be used to highlight the LF band}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL.}

\item{tem}{Boolean, creates a temporal file for the plot. Default is FALSE}

\item{plotHF}{Boolean, plot results form the HF band. Default is TRUE}

\item{plotLF}{Boolean, plot results from the LF band. Default is TRUE}

\item{ratio}{Boolean. Should the LF/HF ratio be plotted? Default is FALSE. Arguments plotHF and plotLF must also be
set to TRUE.}

\item{normalize}{Boolean. Should normalized HRV be plotted? Default is FALSE}

\item{newPlot}{Boolean, generates a new plot without overwriting a previous plot. Default is TRUE}

\item{ylim}{Maximum y axis limit. Default is NULL}

\item{use.ggplot}{Boolean, use methods from \href{https://CRAN.R-project.org/package=ggplot2}{ggplot2} package to plot the results. Default is FALSE}

\item{size.axis}{Percentage of scaling of axis values. Default is 100}

\item{size.labels}{Percentage of scaling of axis labels. Default is 100}

\item{size.title}{Percentage of scaling of plot titles. Default is 100}

\item{d.labels}{Percentage of scaling of distance between axis and labels. Default is 100}
}
\value{
None
}
\description{
Plots the HRV associated to a specific BRS estimation
}
\examples{
Data <- DataSimulation()
Study <- BuildStructure()
Study <- AddAnalysis(Study, name = "Simulation")
Study <- AddDataToAnalysis(Study, 1, Data$RR, Data$SBP, Data$Time, f = 1)
Study <- AnalyzeBRS(Study, 1)
Study <- AddAvgCwtData(Study, 1)


PlotAnalyzedHRV(Study, 1, newPlot = FALSE)
}
\author{
Alvaro Chao-Ecija
}
