% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Individual_BRS.R
\name{PlotBRS}
\alias{PlotBRS}
\title{Plot individual BRS}
\usage{
PlotBRS(
  fun,
  avg = FALSE,
  time_col = "brown",
  HFcolor = "yellow",
  LFcolor = "green",
  time_flags = NULL,
  thr = 0.5,
  use.thr = TRUE,
  nfreqs = 7,
  tem = FALSE,
  newPlot = TRUE,
  title = "data",
  plotHF = TRUE,
  plotLF = TRUE,
  ylim = NULL,
  use.ggplot = FALSE,
  HF = 0.4,
  LF = 0.15,
  VLF = 0.04
)
}
\arguments{
\item{fun}{BRS obtained either by \link[BaroWavelet]{AlphaIndexDWT} or
\link[BaroWavelet]{TransferFunCWT}}

\item{avg}{Boolean. Plot the scale-averaged CWT BRS? Default is FALSE}

\item{time_col}{Color used to highlight a specific time interval. Default is brown}

\item{HFcolor}{Color to be used to highlight the HF band for the average CWT plot}

\item{LFcolor}{Color to be used to highlight the LF band for the average CWT plot}

\item{time_flags}{A vector containing the minimum and maximum limits of a time interval, in minutes.
Default is NULL}

\item{thr}{Coherence threshold to be used for the plot. Default is 0.5}

\item{use.thr}{Boolean, should a coherence threshold be used in the analyses? Default is TRUE.}

\item{nfreqs}{Number of frequencies to be used. Default is 7}

\item{tem}{Boolean, creates a temporal file for the plot. Default is FALSE}

\item{newPlot}{Boolean, generates a new plot without overwriting a previous plot. Default is TRUE}

\item{title}{Part of the title for the plot. Used to indicate the data source on the title. Default is "data"}

\item{plotHF}{Boolean, plot results form the HF band. Default is TRUE}

\item{plotLF}{Boolean, plot results from the LF band. Default is TRUE}

\item{ylim}{Maximum y axis limit. Default is NULL}

\item{use.ggplot}{Boolean, use methods from \href{https://CRAN.R-project.org/package=ggplot2}{ggplot2} package to plot the results. Default is FALSE}

\item{HF}{Maximum limit of the HF band, shown at the plot title. Default is 0.4 Hz}

\item{LF}{Maximum limit of the LF band, shown at the plot title. Default is 0.15 Hz}

\item{VLF}{Maximum limit of the VLF band, shown at the plot title. Default is 0.04 Hz}
}
\value{
None
}
\description{
Plots an already computed BRS from a specific subject
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
AlphaIndex <- AlphaIndexDWT(Data, wv = "d8", error = 0.0005)

PlotBRS(AlphaIndex, plotHF = TRUE, plotLF = TRUE, newPlot = FALSE)
}
\author{
Alvaro Chao-Ecija
}
