% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocessing with RHRV.R
\name{PreprocessData}
\alias{PreprocessData}
\title{Preprocess data with RHRV}
\usage{
PreprocessData(
  data,
  maxSBP = 300,
  minSBP = 30,
  minbpm = 25,
  maxbpm = 200,
  use.RHRV = TRUE
)
}
\arguments{
\item{data}{Cardiovascular data to be processed}

\item{maxSBP}{Maximum value allowed for the SBP}

\item{minSBP}{Minimum value allowed for the SBP}

\item{minbpm}{Minimum bpm allowed}

\item{maxbpm}{Maximum bpm allowed}

\item{use.RHRV}{Boolean. Should RHRV be used in the preprocessing step? Default is TRUE.}
}
\description{
Preprocess data with RHRV
}
\author{
Alvaro Chao-Ecija
}
