% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunShinyApp.R
\name{RunBaroWaveletApp}
\alias{RunBaroWaveletApp}
\title{Run BaroWavelet's ShinyApp}
\usage{
RunBaroWaveletApp()
}
\description{
Launches BaroWaveletApp, BaroWavelet's ShinyApp
}
\examples{
\dontrun{
RunBaroWaveletApp()
}

}
\author{
Alvaro Chao-Ecija
}
