% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test Individual Variables.R
\name{TestIndBRS}
\alias{TestIndBRS}
\title{Test Individual BRS}
\usage{
TestIndBRS(fun, time_flags1, time_flags2, thr = 0.5, use.coherence = TRUE)
}
\arguments{
\item{fun}{BRS obtained either by \link[BaroWavelet]{AlphaIndexDWT} or
\link[BaroWavelet]{TransferFunCWT}}

\item{time_flags1}{A vector containing the minimum and maximum limits for the first time interval, in minutes.}

\item{time_flags2}{A vector containing the minimum and maximum limits for the second time interval, in minutes.}

\item{thr}{Coherence threshold to ensure the reliability of the estimates. Default is 0.5}

\item{use.coherence}{Boolean, should a coherence threshold be used? Default is TRUE}
}
\value{
A vector contaning p-values computed for the HF and LF bands
}
\description{
Significant test to compare two time-domain segments from an individual BRS computation
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)
AlphaIndex <- AlphaIndexDWT(Data, wv = "d8", error = 0.0005)

TestIndBRS(AlphaIndex, c(0, 1.7), c(35, 38))
TestIndBRS(AlphaIndex, c(0, 1.7), c(8, 9.5))
}
\author{
Alvaro Chao-Ecija
}
