% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_Individual_Variables.R
\name{TestIndHRandBP}
\alias{TestIndHRandBP}
\title{Test Individual HR and BP}
\usage{
TestIndHRandBP(data, time_flags1, time_flags2)
}
\arguments{
\item{data}{A matrix with HR and BP recordings, as well as timestamps}

\item{time_flags1}{A vector containing the minimum and maximum limits for the first time interval, in minutes}

\item{time_flags2}{A vector containing the minimum and maximum limits for the second time interval, in minutes}
}
\value{
A vector containing p-values computed for the HR and the SBP
}
\description{
Significant test to compare two time-domain segments in HR and BP recordings
}
\details{
In this function, the Kolmogorov-Smirnov test is used to compare the distributions of
heart rate and blood pressure estimates at the intervals of interest. This procedure 
is based on the strategy used by Rodriguez-Linares et al. to compare time-variable estimates.
}
\examples{
Data <- InterpolateData(DataSimulation(), f = 1)

TestIndHRandBP(Data, c(0, 1.7), c(35, 38))
TestIndHRandBP(Data, c(0, 1.7), c(8, 9.5))
}
\references{
Rodriguez-Linares L, Lado MJ, Vila XA, Mendez AJ, Cuesta P. gHRV: Heart rate variability 
analysis made easy. Comput Methods Programs Biomed. 2014 Aug;116(1):26-38.
}
\author{
Alvaro Chao-Ecija
}
